# PERF: improve perf of conversion to string arrays