@jbrockmendel 2021-09-29
> f"{val}" is faster than str(val)

Looks like this happens bc in the cython version we're using `str` isnt optimized to `PyObject_Str`, but it is in cython 3.0.0.a1 (https://github.com/cython/cython/pull/3478).

Could do this optimization with a note to revert it once we use the newer cython?