@jbrockmendel 2021-09-29
>  Do you have any estimate if the perf. improvement from that 

Not off the top of my head.  Another option would be to use `PyObject_Str` instead of `str` in the interim.

> and when it's released?

Not really clear.  There have been 9 alpha releases since Apr 2020, I don't know what the blocker/timeline is.