@jbrockmendel 2021-09-30
> Using PyObject_Strdidn't give the same speedup as the f-string approach, (732 ms).

well that is surprising.  Any potential explanations in the generated C code?