@topper-123 2021-09-30
> Any potential explanations in the generated C code?

I can look into this when Im at the computer again later today.